Refactor "file_output" module as "state_output"
The state at the initial time is now output.

The output directory is cleared before each simulation start. This prevents
confusion when shorter simulations are run after longer simulations have
already written output.

The output directory, instead of filename, is a directory. This allows more
reliable assumptions about output file names and extensions, and removes
magic templating.

The module should now always be run last, so it is able to capture the
state at the falling edge of each time step. This removes the need to write
a separate "simulation-final.pkl" file.

The file name and class name of this module are now consistent and slightly
more descriptive of the actual functionality.